% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calageInt.R
\name{calageInt}
\alias{calageInt}
\title{Age Intensity Function}
\usage{
calageInt(state, proxel, delta)
}
\description{
This function returns a numeric value indicating
the pending time (in terms of time steps, delta) for a given state change of a proxel.
}
