% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calProb.R
\name{calProb}
\alias{calProb}
\title{Transition Probability Function
This function returns the transition probability under
zero repair time for a given basic event and a proxel.}
\usage{
calProb(proxel, model, state, delta)
}
\description{
Transition Probability Function
This function returns the transition probability under
zero repair time for a given basic event and a proxel.
}
