% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calProb_s.R
\name{calProb_s}
\alias{calProb_s}
\title{Transition Probability Function
This function returns the transition probability under
stochastic repair time for a given basic event and a proxel.}
\usage{
calProb_s(proxel, model, state, delta)
}
\description{
Transition Probability Function
This function returns the transition probability under
stochastic repair time for a given basic event and a proxel.
}
