% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_transition.R
\name{next_transition}
\alias{next_transition}
\title{State and the cost of the next transition}
\usage{
next_transition(state, K, t, model, alpha, W)
}
\arguments{
\item{state}{a string value for the state}

\item{K}{a numeric value as a decision variable (between 1 and N)}

\item{t}{a numeric value as current time}

\item{model}{a list containing states, transition matrix, distributions and their parameters}

\item{alpha}{a numeric value as a decision variable (between 0 and W)}

\item{W}{a numeric value as the warranty period}
}
\value{
a list containing the next state and the cost of the next transition
}
\description{
For a given basic event with stochastic repair time, this function
calculates the next state of the system with the associated cost.
}
