% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextLevel_s.R
\name{nextLevel_s}
\alias{nextLevel_s}
\title{Proxels of the next time step}
\usage{
nextLevel_s(proxel, model, delta, t, alpha, W, K)
}
\arguments{
\item{proxel}{a data frame containing a state, age intensity and a probability.}

\item{model}{a list containing states, transition matrix, distributions and their parameters}

\item{delta}{a numeric value as time step}

\item{t}{a numeric value as current time}

\item{alpha}{a numeric value as a decision variable (between 0 and W)}

\item{W}{a numeric value as the warranty period}

\item{K}{a numeric value as a decision variable (between 1 and N)}
}
\value{
a data frame where each row is a proxel
}
\description{
For a given basic event with stochastic repair time, and a proxel, this function
calculates all the possible proxels for the next time step.
}
