% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Estimates the model parameters from the data}
\usage{
parameters(data, use_estimated = TRUE, cutpoint, policy)
}
\arguments{
\item{data}{a data frame containing the data set}

\item{use_estimated}{a logic value indicating whether to use estimated states or not.}

\item{cutpoint}{a numeric value as the cutpoint}

\item{policy}{a numeric value as the warranty policy (1 or 2)}
}
\value{
a list of parameters
}
\description{
For a given dataset and a warranty policy, this function
calculates all the model parameters.
}
