% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreData.R
\name{PreData}
\alias{PreData}
\title{Pre processes a data set}
\usage{
PreData(data, W)
}
\arguments{
\item{data}{a data frame containing the data set}

\item{W}{a numeric value as the warranty period}
}
\value{
a data frame processed data
}
\description{
For a given data set, this function pre processes the data for .
}
