% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxel.R
\name{Proxel}
\alias{Proxel}
\title{Instantaneous Unavailability Vector}
\usage{
Proxel(state, W, model, delta, tol, alpha, K)
}
\arguments{
\item{state}{a string value for the state}

\item{W}{a numeric value as the warranty period}

\item{model}{a list containing states, transition matrix, distributions and their parameters}

\item{delta}{a numeric value as time step}

\item{tol}{a numeric value for the tolerance level}

\item{alpha}{a numeric value as a decision variable (between 0 and W)}

\item{K}{a numeric value as a decision variable (between 1 and N)}
}
\value{
a numeric vector of warranty costs
}
\description{
This function calculates the isntantaneous unavailablity values
of a product with zero repair time, along with the average warranty cost.
}
