% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData.R
\name{SimData}
\alias{SimData}
\title{Simulates warranty logs data}
\usage{
SimData(alpha, K, W, model, SimSize, path)
}
\arguments{
\item{alpha}{a numeric value as a decision variable (between 0 and W)}

\item{K}{a numeric value as a decision variable (between 1 and N)}

\item{W}{a numeric value as the warranty period}

\item{model}{a list containing states, transition matrix, distributions and their parameters}

\item{SimSize}{a numeric value as the simulation size}

\item{path}{a string value to save the simulated data}
}
\value{
a data frame containing warranty logs data
}
\description{
For a given set of parameters, simulates warranty logs data
}
