% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateEstimation.R
\name{StateEstimation}
\alias{StateEstimation}
\title{Estimates the degradation state of the product using cost of repair/replace}
\usage{
StateEstimation(data, policy = 2)
}
\arguments{
\item{data}{a data frame containing warranty logs}

\item{policy}{a numeric value as the warranty policy (1 or 2)}
}
\value{
a list containing the plots of the warranty logs,
a data set with estimated states, accuracy and the cutpoints
}
\description{
For a given data set of warranty logs, estimates
the degradation state of the product from its cost of repair/replace
}
